Move examples to own files
