+++ Tests/MountebankSwiftTests/ExampleData/Stub/Response/Fault+Examples.swift
@@ -1,0 +1,16 @@
+import Foundation
+import MountebankSwift
+
+extension Fault {
+    enum Examples {
+        static let connectionResetByPeer = Example(
+            value: Fault.connectionResetByPeer,
+            json: "CONNECTION_RESET_BY_PEER"
+        )
+
+        static let randomDataThenClose = Example(
+            value: Fault.randomDataThenClose,
+            json: "RANDOM_DATA_THEN_CLOSE"
+        )
+    }
+}
