+++ Tests/MountebankSwiftTests/ExampleData/Stub/Response/Inject+Examples.swift
@@ -1,0 +1,13 @@
+import Foundation
+import MountebankSwift
+
+extension Inject {
+    enum Examples {
+        static let injectBody = Example(
+            value: Inject(
+                "(config) => { return { \"body\": \"hello world\" }; }"
+            ),
+            json: "(config) => { return { \"body\": \"hello world\" }; }"
+        )
+    }
+}
